# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0",
    "description": "A small CRUD API for a to-do list. Tasks live in memory only, so the list resets every time the server restarts."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/": {
      "get": {
        "summary": "Describe this API",
        "responses": {
          "200": {
            "description": "Name, version and available endpoints",
            "content": {
              "application/json": {
                "example": { "name": "Task API", "version": "1.0", "endpoints": ["/tasks"] }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Check the server is alive",
        "responses": {
          "200": {
            "description": "The server is up",
            "content": { "application/json": { "example": { "status": "ok" } } }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Count total, done and open tasks",
        "responses": {
          "200": {
            "description": "A summary of the list",
            "content": {
              "application/json": { "example": { "total": 3, "done": 1, "open": 2 } }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Restore the 3 example tasks",
        "responses": {
          "200": {
            "description": "The freshly seeded list",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks, optionally filtered",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "required": false,
            "description": "Keep only finished (true) or unfinished (false) tasks",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Keep only tasks whose title contains this text",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The matching tasks",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } }
              }
            }
          },
          "400": {
            "description": "Invalid value for the done filter",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" },
              "example": { "title": "Buy milk" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          },
          "400": {
            "description": "Title missing or empty",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "The id of the task",
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "summary": "Get one task by id",
        "responses": {
          "200": {
            "description": "The requested task",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          },
          "404": {
            "description": "No task with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "put": {
        "summary": "Update a task's title and/or done flag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskUpdate" },
              "example": { "title": "Buy oat milk", "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          },
          "400": {
            "description": "Empty or invalid body",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "No task with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "responses": {
          "204": { "description": "Deleted — no content returned" },
          "404": {
            "description": "No task with that id",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Buy milk" },
          "done": { "type": "boolean", "example": false }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": { "title": { "type": "string", "example": "Buy milk" } }
      },
      "TaskUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Buy oat milk" },
          "done": { "type": "boolean", "example": true }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string", "example": "Task 99 not found" } }
      }
    }
  }
}